Add Resource binding test
A new test that asserts that the server closes the connection with an error, when a stanza is sent prior to resource binding (as defined by RFC 3620, section 7.1).

Smack doesn't offer an API that can be utilized to send a stanza between authentication and resource binding. For that reason, this test sends stanzas prior to authentiation (instead of post auth, pre bind).