Have project generate an executable jar
To prevent projects from having to compile this project when they want to use it, an executable should be published.

In order to be able to publish an executable, such an executable must first be generated. This commit ensures that a stand-alone executable jar is generated during the maven 'package' phase.